+++ services/frontend/src/app/layout/common/digital-service-table/digital-service-table.component.ts
@@ -1,15 +1,18 @@
 import { Component, computed, EventEmitter, inject, Input, Output } from "@angular/core";
 import { TranslateService } from "@ngx-translate/core";
+import { ConfirmationService } from "primeng/api";
 import { UserService } from "src/app/core/service/business/user.service";
 import { GlobalStoreService } from "src/app/core/store/global.store";
 
 @Component({
     selector: "app-digital-service-table",
     templateUrl: "./digital-service-table.component.html",
+    providers: [ConfirmationService],
 })
 export class DigitalServiceTableComponent {
     protected userService = inject(UserService);
     protected translate = inject(TranslateService);
+    private readonly confirmationService = inject(ConfirmationService);
     private readonly globalStoreService = inject(GlobalStoreService);
 
     @Input() data: any[] = [];
@@ -20,6 +23,7 @@ export class DigitalServiceTableComponent {
     @Input() headerFields: string[] = [];
     @Input() showId = true;
     @Input() addButtonId = "add-button";
+    @Input() isVM = false;
 
     @Output() sidebar: EventEmitter<boolean> = new EventEmitter();
     @Output() resetItem: EventEmitter<boolean> = new EventEmitter();
@@ -49,5 +53,23 @@ export class DigitalServiceTableComponent {
             ...item,
             index,
         });
+    }
+
+    confirmDelete(event: Event, item: any, index: number) {
+        if (this.isVM) {
+            this.doDeleteItem(item, index);
+        } else {
+            this.confirmationService.confirm({
+                closeOnEscape: true,
+                target: event.target as EventTarget,
+                acceptLabel: this.translate.instant("common.yes"),
+                rejectLabel: this.translate.instant("common.no"),
+                message: this.translate.instant("common.resourcesPopup.delete-question"),
+                icon: "pi pi-exclamation-triangle",
+                accept: () => {
+                    this.doDeleteItem(item, index);
+                },
+            });
+        }
     }
 }